Added Dockerfile to run api code, added .env file, changed requirements.txt
Got rid of json from requirements.txt and added in package versions for pandas and requests.
Made a .env file with my API key so that Docker can access it when building and running the image.
Created a Dockerfile to run the API script - made some changes with the path.